+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 1.0.15 - 25 Apr 2024
+# 1.0.15 - 27 Apr 2024
 Feature:
 - add `redirect` function to `Context`